Update Mastermind.java
While the initial challenge did not say I had to give the user as much information as this update does, I went ahead and did it anyway. It was as easy as adding a new array and iterating through the array. I also just used another int so that the correct guess counter always outputs an accurate value but can also be safely reset to 0 before going through any checks or modifications.